Check whether there are healthy backend available
Currently when there are no backend available in pool Mantrid is
throwing general exception. We want to handle it in more civilized
way so nginx can catch a problem and fallback to other grid
clusters.

This fix is needed for az-separation, as when setting up new cluster
it will not have any backends set in mantrid. As workers connect
only to nginx->mantrid correlated to their grid, we don't want them
to throw errors and bother anyone, so let's just let nginx to
connect to other, healthy grid cluster.